Make a wrapper script for ld which adds the hugepage linker script in the correct place.  Use this in preferance to directly manipulating the link command line.
This will make life easier if we ever need to add more linker options than
just the script.  It's also handy for things like XLC/XLF which add their
own libraries after the given linker options.  The script (and hence
the link of libhugetblfs) *must* go last, so that it's constructor is called
as early as possible.