[PATCH] libhugetlbfs: Use gcc -B when invoking ld wrapper script
On SLES9, at least, gcc is configured to call ld via an absolute path.
Therefore, simply arranging for our ld wrapper to appear first in the
PATH will not get it called by gcc.  Luckily the option -B allows us to
specify a place to look for commands invoked by the compiler.  Use that
instead so libhugetlbfs can be compiled again.