Make the chunk-overcommit test more selective.
This patch improves the chunk-overcommit test.  Previously it would assume
any apparent overcommit was bad, and FAIL accordingly.  It now goes on to test
if the overcommit actually results in a SIGKILL (or other failure) when
you attempt to instiate all the allocated pages.  This means this test will
not give spurious failures if we ever get a kernel which automatically tries
to make more hugepages, or can swap hugepages or can otherwise cope with an
overcommit situation.

Also added a test rationale comment.