copy only p_filesz bytes from original segment into huge pages
For ELF segments with a significantly large area of uninitialized data (memsz is much larger than filesz) it is incredibly inefficient to copy data beyond filesz bytes since it should be all zeroes anyway.

Making this optimization requires small changes to the linker scripts that remap the data segment into huge pages.  The plt section must be padded with a piece of static data to force it into the filesz (on-disk) portion of the executable.  This ensures the plt actually gets copied during remapping.