Use mlock to reserve newly allocated hugepages in hugetlb_morecore.
Since MAP_PRIVATE allocations are full overcommit, the mmap call can succeed
even if there aren't enough huge pages to satisfy the mapping.  Thus fallback
to normal pages is not currently possible.  Use mlock to instantiate the newly
allocated area and if that fails, release it and initiate fallback to normal
pages.