Don't kill the daemon if shared_file allocation fails
Currently, we too aggressively kill the daemon, just because we failed
to create a new hugetlbfs file. Legitimately, the daemon should stick
around, as sharing of existing segments can continue; and we might get
hugepages again eventually.  Make sure the error path back from
hugetlbfs_find_path() is correct, as well (which is how this problem was
discovered).

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>