Add verbosity command line parameter to hugetlbd
As part of the goal of minimizing setting environment variables, I added
a flag to hugetlbd to specify the verbosity of logging, as opposed to
setting HUGETLB_VERBOSE. I also went ahead and cleaned up the usage()
statement. The Makefile change indicates that we no longer depend on
debug.o for the debugging symbols. We still need hugetutils.o, however,
for the creation of unique hugetlbfs files. Probably applies best on top
of my immediately previous change to hugetlbd.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>