build: Fix compile when __syscall macros aren't defined
Alright, given our __syscall_return horkiness and it's imminent removal,
I'm trying to figure out a good way to remove all usage of it from
libhugetlbfs.

As a first step, I've got the following workaround building on i386 both
in a Rawhide and not in a Rawhide environment (in the absence and
presence of __syscall_return respectively). It also builds the x86_64
version fine still.

Thoughts? I would really like to just get rid of the macros completely,
as we should not depend on a kernel include (asm/unistd.h via
linux/unistd.h), but I'm not quite sure how best to do so. The patch
looks a lot worse than it is, it just moves the code around.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>