elflink: Smarter copysize detection v2
When a program starts up, its BSS should be all zeroes.  Unfortunately, since
libhugetlbfs depends on glibc, that library is loaded first.  It's constructors
get run, and those constructors may initialize some global variables in the
BSS.  If this happens, and libhugetlbfs does not copy that initialized data,
the application will almost certainly crash.

The patch introduces logic that utilizes information in the dynamic section to
look up global symbols in the bss.  It then ensures that all of these symbols
get copied into the hugepage mappings as well.  In the future it will be
possible to further trim down this extra copy by only copying glibc symbols.

This should prevent users from needing to use the HUGETLB_MINIMAL_COPY=no
behavior when applications crash upon startup.  This patch is based Adam
Litke's copysize patch and was developed in collaboration with him.

Signed-off-by: Steve Fox <drfickle@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>