elflink: Disable segment remapping when libhugetlbfs is preloaded
One of our testers has been hitting the following problem:

When running an application linked against a libhugetlbfs-provided
linker script, LD_PRELOAD=libhugetlbfs.so is accidentally set.  The
library incorrectly tries to perform remapping, which results in an
application segfault.

We must add extra checking to handle this case since our normal method
of disabling segment remapping won't work in this scenario.  This patch
checks for "libhugetlbfs" in the LD_PRELOAD environment variable.  If
found, remapping is disabled and the user is warned.  With the addition
of an additional environment check, I felt it was time to move them into
a separate function.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: Nishanth Aravamudan <nacc@us.ibm.com>