morecore: Disable heap trimming
Since we may need to align the base of the heap up to a hugepage boundary, our
morecore implementation breaks brk() semantics.  It doesn't become a problem
until glibc tries to trim the heap beyond a certain point.  Since we're not
freeing huge pages in response to glibc trimming requests anyway, disable
trimming altogether to avoid this problem.  We may be able to fix this issue in
the future but it will require a fair amount of development and a decent
increase in code complexity.

Signed-off-by: Adam Litke <agl@us.ibm.com>