install: Fix make install error when a stale symlink is present
Since $BINDIR is "owned" by libhugetlbfs, any existing symlinks in that
directory can be assumed to be stale.  It is therefore safe to create the ld
symlink on top of any old one which might be in place.  This eliminates a make
install error when re-installing to the same location.

Signed-off-by: Adam Litke <agl@us.ibm.com>