tests: Select number of threads for linkshare based on free huge pages
Description: Since we now pad the BSS of relinked binaries, we now
require a larger number of hugepages than before, even if most of them
are unused. This leads to issues with the linkshare testcase, as it
spawns a fixed number of threads, all of which will consume hugepages
and eventually lead to a ENOMEM (in hugepages) condition. Modify the
testcase invocation to spawn a number of threads relative to the number
of free hugepages (even if the BSS padding is fixed differently, this is
a reasonable thing to do). Also modify the linkshare testcase to do
nothing if no threads are requested (which will now occur if the number
of hugepages free in the system is 0).

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>