build: Clarify use of $(DESTDIR)
Commit 70d6e658872b9f0d1b1688e0ef7cc4c711c9161e fixed a make bug due
to confusion as to whether $(BINDIR), $(LIBDIR32) and so forth
referred to the final installed directory of the files in question, or
to where they'll be placed by make install - with DESTDIR used for
buildroots, they may be different.

However that commit led to the semantics being different for
LDSCRIPTDIR from the other variables, which is confusing.  This patch
clarifies the situation: $(BINDIR) etc. always refer to the final
installed directory, $(DESTDIR) is only prepended in the final install
commands.  IIRC, this is also the way DESTDIR normally works in
automake projects and the like.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Acked-by: Steve Fox <drfickle@us.ibm.com>