tests: Simplify and speed up the mlock testcase.
At present the mlock test case attempts to mlock() all available
hugepages (as passed in on the command line).  This isn't particularly
more useful than just testing locking a single page.  This patch
reduces it to testing just a single page (with each set of mapping
options).

In addition, we were only testing actual mlock() calls on files opened
with MAP_LOCKED, which are supposed to be mlock()ed anyway, rather an
oversight.  This patch corrects this also.
 What would be another useful test is to verify that mlock()ing really
does reserve the page in question, i.e. that a page mlock()ed but not
touched will always instantiate without fault later, even if all other
hugepages on the system have been consumed.  That's another patch for
another day.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Acked-by: Nishanth Aravamudan <nacc@us.ibm.com>