install: Always install linker scripts for all targets
This patch changes the install target to always install the linker
scripts for all supported platforms, instead of just the ones think
will be used for the current architecture.  This avoids some
complications in corner cases as to which scripts will be needed, and
could also make life more convenient for cross-compiles.  The linker
scripts are small, so there's no strong reason to avoid installing
them.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Acked-by: Steve Fox <drfickle@us.ibm.com>