ldscripts: Missing bss section placement statement
This appears to be a longstanding bug in the ppc32 .xB linker script.  When the
BSS was moved into its own htlb segment, we lost the explicit placement of the
.bss* and .gnu.linkouce.b.* sections.  Default ld behavior was placing them
implicitly in the htlb segment anyway so we didn't notice the problem until
commit c3f4d10a80d6ad8bb37ef6f80e70c3c7d778117a "ldscripts: [powerpc] Align the
end of the BSS to a segment boundary".  After this change, the implicit
behavior was to place the ommitted sections after the 256M alignment (which
defeats the purpose of the ALIGN in the first place).

The default system linker scripts (on SLES10 at least) contain the line added
by this patch, which further supports my case that it should be there.

Signed-off-by: Adam Litke <agl@us.ibm.com>