hugetlbd: Move socket to /var/run
There are at least two fundamental problems with placing the hugetlbd
socket in /tmp.  First, it could be deleted at any time by a /tmp
reaping daemon.  Second, the lax permissions of /tmp could allow an
unprivileged application to impersonate the daemon by commandeering the
socket file.  Move the socket to /var/run to avoid this problem.  Note
that on most systems, hugetlbd will need to be run as root in order to
create the socket in the more secure /var/run directory.

Signed-off-by: Adam Litke <agl@us.ibm.com>