tests: Always use two threads for the linkshare tests
Commit 1b5e42ad16ab3df69283f114a7822e3e7fbc238a changed the run_tests
script to change the number of threads for the linkshare test based on
the available hugepages.

This was done to allow the test to run on systems without vast
quantities of hugepages, but it's not the best approach.  For one
thing it will happily reduce the number of threads to 1, which rather
defeats the purpose of a sharing test.  It also makes unwarranted
assumptions about how much memory the given test program ($baseprog)
needs.

The real problem was that originally we were arbitrarily running 10
threads, for no particular reason.  In accordance with the principle
of minimal testcases (at least for the "functional" section of the
testsuite), this patch removes the adaptive number of threads logic
and just runs 2 threads - the minimum number we need to have a
meaningful test of the elflink sharing code.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>