tests: use the syscall library function to access gettid()
Sadly, none of the standard headers include a proper prototype for the
gettid() system call, but we need it for the private version of
alloc-instantiate-race.  Our current method of getting it - invoking
the _syscall0 macro doesn't work on all systems.  This patch
implements a more robust approach, using the syscall() library
function.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>