build: Use literal tab instead of '\t'
Currently, our less-verbose mode for make (adapted from Kbuild), uses
an "echo -e \t" to print a tab for spacing.  However, this doesn't
work correctly with all versions of echo.  (for example, the recent
change in Ubuntu edgy to a non-bash /bin/sh causes every line to be
preceded with a bogus literal "-e").

This patch replaces the use of "echo -e" with an explicit tab,
included literally in the Makefile.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>