elflink: Remove sharing daemon
Ok.  I tend to think it will get its testing fastest if we merge into
the git tree now.  Again, we can branch from 1.0.1 if we need to make
more maintenance releases for Red Hat which shouldn't include the new
code.

So, here is the latest patch, revised according to your comments.  I
haven't yet implemented the check-permissions after hugepage file
open() trick, but I will look at that later.

Abolish the hugepage sharing daemon, hugetlbd

At present, sharing of hugepages between multiple invocations of a
prgram using the elflink features is accomplished using a daemon.  The
daemon keeps track of current hugepage files with segment data, and
supplies their names to processes requesting it.

The daemon, however, adds considerable complexity, both to the
implementation, and to the user or administrator (who must ensure that
the daemon is properly started before sharing is attempted).  Worse,
the way the daemon operates introduces a number of serious security
flaws, which can't easily be addressed without even more complexity
for extra credential checks and the like.

But, the daemon is unecessary.  Instead, we can simply use a standard
naming convention so that processes can find pre-prepared, shareable
files in hugetlbfs with the correct segment data.  This patch,
therefore, removes the daemon replacing it just such an approach.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>