tests: Add some truncation race tests
Some kernels have a bug where the test of the faulting offset against
i_size is misplaced.  This has two bad consequences:
        - if there is a fault on an address past the end of file, and
the pool of hugepages is also exhausted, we can get an OOM, SIGKILLing
the process, instead of the SIGBUS we should get.
        - a fault on an address past the end of file can cause the
count of reserved hugepages to wrap around "negative" (actually to a
huge value, since it's unsigned)

This patch adds testcases for both problems.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>