elflink: Restore unmapped_abort debugging info
On at least 2 of my ppc32 systems this patch also appears to trigger
assertion failures in ld.  However, despite the assert failures, it
still seems to generate correct output.  The asserts also appear to be
fixed in mainline binutils, so I'm inclined to ignore this problem.

At present, we attempt to use the _syscallXX() macros to produce some
debugging output if things go wrong in the elflink code while the PLT
is unmapped (and hence can't use normal libc functions).  This is
fragile: the _syscall() macros aren't always available, and don't
always work properly even when they are.

This patch changes this section of code to instead use the syscall()
library function.  Some linker / makefile tricks are then used to
ensure that we link statically with the syscall() symbol from libc,
actually importing a copy into libhugetlbfs.  Other libc library calls
remain dynamically resolved as before.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>