Testcase for misaligned offset maps
The kernel (correctly) returns EINVAL if you attempt a hugepage
mapping with a file offset which is not hugepage aligned.  However,
current kernels have a bug in the corresponding backout path, which
attempts to use the normal unmap_region() on the (half-constructed)
hugepage VMA.  On ppc64, where the area could be in a region already
set aside for hugepages, this can cause bad_pud() to be triggered with
consequent nastiness.

This patch adds a testcase to libhugetlbfs to catch this bug.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>