Make ar less verbose during build
The default value of ARFLAGS provided by make is "rv", which tells ar
to be verbose, listing each object it adds to the archive.  When
building without V=1, this interrupts flow of brief messages from
VECHO, making genuine warnings slightly less prominent.

Thus, this patch alters the ARFLAGS when in non-verbose build mode to
"rc" which turns of ar's verbosity, and also suppresses ar's
message/warning that it is creating a new archive.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>