tests: Fix zero_filesize_segment testcase
This testcase is implemented as a linker script which can be "compiled"
like a program.  As such, the caveats with .note.ABI-tag and .note.SuSE
in our other linker scripts apply here as well.  Add explicit placement
of these sections so ld's implicit guessing doesn't get it wrong.

At least it wasn't a kernel regression ;)  Please apply.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>