Fix truncate_above_4GB test for wider range of pagesizes
On powerpc with 4K base page size, the truncate_above_4GB test fails
in an uninteresting way (i.e. not because of the bug it was designed
to catch).  This because the 'buggy_offset' it calculates is not
hugepage aligned, so the map at this offset (correctly) fails.

This patch rounds the offset up to a hugepage size to avoid this.
According to Hugh Dickins who provided the initial testcase, this
still ought to catch the original problem.  It's more-or-less
impossible to test, though, since the bug only ever showed up on
32-bit archs, i.e. on i386 and on i386 the pagesizes are such that the
calculation doesn't result in a misaligned offset.

While we're at it, this patch adds some more detailed verbose_printf()
statements to make tracking down test failures easier.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>