Testcase for kernel bug attempting brk() into a hugepage area
Current kernels have a bug in the brk() codepath.  This path, unlike
the mmap() path, doesn't check if the newly allocated region
intersects a hugepage dedicated region.  This means it can create a
normalpage VMA which extends into a hugepage-only address region (at
least on powerpc, which has such regions).  This can easily lead to
later oopses or other nastiness.

This patch adds a testcase to catch this bug (triggering an oops on
powerpc).

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>