Clean up hugepages before/after linkshare tests
After the daemon removal, there was a change in the side-effects of
running `make check` relative to the linkshare testcases, as hugepages
would be left consumed on the system, that were not before the test ran.
`rm -rf` the files we create (we assume the testsuite is the only
hugepage consumer on the system in general, when run), to free them back
up after running. Run the elfshare test twice now, once with this
cleaning up of hugepages and once without.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>