Update test rationale comments
This went to the list before, but it was a reply on a different
thread, so probably got missed.  Please apply.

This patch adds some extra details to comments in a number of
testcases, giving useful information about the rationale and origins
of the test.  In particular it adds a prominent warning on testcases
which include "magic" offsets and addresses chosen carefully to
trigger particular subtle kernel bugs.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>