Add and use a helper program for the hugetlbfs mountpoint
It was noted recently that checking for a hugetlbfs mountpoint in the
run_tests.sh can lead to a few issues (spaces in /proc/mounts, for
instance), which are hard to work around. Instead, make use of the
library's utilities via a small helper program which should provide
greater safety.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>