tests: Add direct-io testcase
By opening a normal file with O_DIRECT, subsequent reads from and writes
to that file will use the kernel's Direct-IO layer to reduce the amount
of copying to/from user space.  If the buffer used in that read/write is
located within a hugetlb page mapping, direct-IO to or from huge pages
can occur.  This testcase checks to make sure this scenario works
correctly.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>