make sure to use newlines in debugging/error paths
Some of our error and debugging paths are missing newlines, which makes
for confusing output when things are misconfigured. Make sure all ERROR
and WARNING in the library is properly terminated. An alternative fix
would be to terminate the macros themselves with newlines, but this
would be more invasive and conflicts with another patch I would like
merged. I will revisit this post 1.0.2.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>