Remove newlines in some debugging to save space
I noticed that we currently put the "done"s from the segment remapping
debugging on their own line. But I'm not sure I see the value, sine it
seems rather out of place, given that we specifically use DEBUG_CONT().
I think it looks cleaner (and saves up to two lines of screen space) to
not newline-terminate the previous output.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>