remove sharing of writable segments
Address space randomization on x86_64 causes the sharing testcase (and
segment sharing in general) to fail for HUGETLB_SHARE=2 (sharing all
segments). However, this is only one of potentially several ways that
this feature might lead to problems. Remove the ability to specify
sharing of all segments, by warning the user if HUGETLB_SHARE=2 and
disabling sharing. Update the HOWTO and the linkshare testcase as well.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>