readd the mlock() call to morecore
We currently run into problems with the libhugetlbfs malloc
implementation if hugepages are available when we mmap, but not
available when we touch them. The application is killed, even though
there were sufficient hugepages at some point in its lifetime. We have
code which uses mlock() to bring in all the hugepages in our heap, but
it was commented-out because of bad NUMA performance (all hugepages
would be allocated node-local). Bring back that code, as the kernel-side
issues with NUMA and hugepages have been resolved (interleave works
again), and add a comment indicating NUMA users should use numactl to
specify the hugepage allocation policy.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>