Fix extracopysize detection algorithm
The current extracopysize detection algorithm does not include weak
objects in the copy window.  These are symbols marked with 'W' and 'V'
in nm output.  These can in fact be initialized global variables (just
like STB_GLOBAL symbols) and should be included in the extracopy window.

The following patch adds STB_WEAK symbols to the extracopy window.
Since the if statement is getting a bit large and out of hand, move the
logic out of the function and into its own macro "keep_symbol".

This patch resolves an eon (speccpu2000 benchmark) failure.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>