New testcase: stack_grow_into_huge
Sonny Rao discovered a bug which affects powerpc machines.  If the stack
is allowed to grow into a segment that has been converted over to huge
pages, current kernels will try to instantiate normal pages into a huge
page-only segment.  This test case reproduces this behavior for 32bit
and 64bit binaries.

Since there is no upstream fix for this bug yet, I don't think it's wise
to run it all the time and needlessly render machines unstable.
However, we don't want to forget that it's not being run by default
either.  This patch also introduces a SKIP mode to run_tests.sh.  It
will cause the named test to be skipped and a message printed that
indicates that a test was skipped.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>