elflink: drop hugepage cached pages for writable segments
We currently use extra hugepages for writable segments because of the
first MAP_SHARED mmap() which stays resident in the page cache.  Use a
helper function to force a COW for the filesz portion, as well as the
extracopy area, of writable segments and then fadvise() to drop the page
cache pages while keeping our PRIVATE mapping. This is mutually
exclusive to segment sharing, but is also orthogonal in code because we
only allow sharing of read-only segments.  Also, if the minimal_copy
algorithm is disabled, we disable this optimization.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>