hugeutils: remove hugetlbfs_vaddr_granularity in favor of slice functions
hugetlbfs_vaddr_granularity() currently does not take a parameter, even
though the granularity at which hugepages can be used varies by address
on ppc64 (and perhaps other archs). Taking a queue from Ben
Herrenschmidt's kernel slice patches for Cell add similar functions and
macros for hugetlbfs_slice_start() and hugetlbfs_slice_end() which,
given an unsigned long address, will indicate where the enclosing slice
starts and ends, inclusively (that is, the start and end are part of the
slice). This has very clean and useful semantics compared to the
granularity() function. This is also useful for the partial_remap code,
as it would (effectively) never run on ppc64 otherwise (would need a 1TB
or larger segment, without relinking, which is unlikely). Thanks to
David Gibson for several rounds of review and one compile-fix.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: David Gibson <dwg@au1.ibm.com>