elflink: add extra debugging for binaries that may run out of address space
We've found that on Power some 32-bit binaries, when relinked via
libhugetlbfs, run out of address space. Add some heuristics to print out
the total memory size of all hugepage segments, and the memory size of
the largest segment. This information, along with the available address
space to applications, can be useful in determining if the segments are
too large for the address space.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>