Fix SEGVs in task-size-overrun test
Currently, the task-size-overrun tests determines TASK_SIZE by
searching the address space from the top of the stack mapping, until
it gets an mmap() error indicating that it's run out of address space.

For this to work safely, the stack must be the highest existing
mapping, which is not always true for recent kernels, at least on i386
and x86_64.

This patch fixes the test by having it search from the top of the
highest existing mapping, whether or not that is the stack.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>