Add per-target-arch syscall stubs to the library
Here's a revised version of my patch adding per-target syscall stubs
to the library.  This is now tested on x86_64.

This patch adds a per-target module to the library, which can be used
to hold any arch or target dependent code.  It uses this module to
provide a per-arch re-implementation of syscall(), which is used to
(finally) provide reliable error messages from elflink.c if they occur
while the program segments are unmapped.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>