Fix the icache-hygeine testcase for i386 and x86_64
The exact CPU behaviour on i386 and x86_64 in the PASS case for the
icache-hygeine test depends on the details of the compiler's code
generation.  Previously, we only tested for one of the possible
behaviours, meaning the test could SEGV depending on the exact
compiler version and options.

This patch makes the test cover the other case as well, making it much
more robust.  It also adds some extra comments explaining more clearly
what happens in the PASS case.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>