tests: use appropriate return type for gethugepagesize()
While running `make func` on ia64, I found that
truncate_sigbus_versus_oom was adding a negative value to a pointer and
causing a SEGV. This was because hpage_size was of type int, even though
the return type of gethugepagesize() is long. The multiplication was
truncated and led to negative values. Fix this throughout the tests,
just in case, by making hpage_size always a long.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>