Makefiles: fix build on ia64 by expressing dependencies
Make the elflink.o and direct_syscall targets depend on ELF32 or ELF64
being defined.
IA64 will not be able to pass tests that rely on MAP_FIXED semantics for
hugepages, which the LIB_TESTS_64 tests all do. So don't build them on
that arch (requires an extra export). run_tests.sh says no such
file, but won't FAIL now.
Make LDSCRIPT_TESTS depend on the existence of ELF32 or ELF64 (that is,
on the existence of linker scripts), just like the relinking tests.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>