tests: update for ia64
truncate_above_4GB: Must be built like it is on powerpc due to address
space restrictions on hugepages.
brk_near_huge: add next_chunk case for ia64 to express that IA64 huge
page mappings start at 0x8000000000000000UL (2^63).
icache-hygeine: IA64 also gets a SIGILL in the icache-hygeiene test, so
add it to the powerpc conditional and update the comment.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>