morecore: mlock() failure is non-fatal
Kernels before 2.6.17 did not properly support mlock() on hugepages (for
instance, the SLES10 kernel). But we currently fail out of the morecore
code, if mlock fails (since 1.1). This error may not be fatal -- and if
it is (that is, we're racing with another process for bringing
hugepages in), then we're going to fail anyways.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>