Revert ae76f97ae76f97... hugeutils: remove hugetlbfs_vaddr_granularity in favor of slice functions
After several rounds of discussion with David Gibson, I believe the
semantics of the slice functions is too hard to get right (esp. on ia64)
to make it an exported and useful interface. morecore and brk_near_huge
are fine with using either the granularity function, or their own custom
version, respectively.

The real consumer of these new interfaces will be the partial_remap
code, but in that case, we'll put all the functions local to elflink.c.

This reverts commit ae76f97ba81aad38e7a2064bfe63e919232c7053.