hugeutils: move hugetlbfs_vaddr_granularity
The semantics of the granularity function are such that it is only
useful in morecore itself (and it's only used there, for that same
reason). Move the definition to morecore.c and remove the declaration
from hugetlbfs.h (it shouldn't have been there anyways, since it isn't
visible outside of the lib).

Also, add an ia64 to the now morecore-specific function.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>