elflink: rework plt detection
We currently emit LONG(0) into the .plt section to make it appear in the filesz
(on-disk) portion of the data/bss segment. This is a problem, however, for
powerpc64, where the ABI specifies that the .plt section is NOBITS (x86 and
x86_64 mark the .plt PROGBITS). Given that the program is being relinked to
begin with, however, it seems logical to simply add a tag, similar to
__executable_start, to make finding the "libhuge" filesz easy to find. Do
exactly this, via __libhuge_filesz. We continue to use the extracopy algorithm,
but move the extracopy end out to __libhuge_filesz if it is after the detected
end.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>