elflink: remove redundant code
The checking of fds >0 is not needed because it is already done earlier
in the function (find_or_prepare_shared_file) and control leaves the
function before there is any chance for it to hit the fail: label.

Note: if we go back and fix the FIXME to add a timeout to the loop,
we'll want to readd the close(fds) to prevent a leak.

Signed-off-by: Jon Tollefson <kniht@linux.vnet.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>